requirements.txt
crewai>=0.28.0
langchain-openai>=0.1.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
pandas>=2.0.0
pydantic>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
